Ethereum test fix, move to one test file.
